+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.5.0 (...)
+- switch to Bearer authentication from Basic Authentication to improve
+  performance (*BREAKING CONFIG*)
+
 ## 2.4.2 (2016-02-22)
 - redo input validation, fix some small bugs and only do input validation 
   in the modules